requirements.txt
# --- Framework principal ---
Flask==3.1.1

# --- Extensiones de Flask ---
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.20
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3

# --- Dependencias de Flask ---
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0

# --- Otros ---
boto3
python-dotenv